+++ NEWS.md
@@ -3,6 +3,7 @@
 - Support use of RasterStack weights in exact_extract named summary operations
 - Support use of weights in exact_extract R summary functions
 - Support SpatialPolygons and SpatialPolygonsDataFrame inputs
+- Support exact_resample with mode, minority, variety, median, quantiles
 
 # version 0.5.1